fix(gdn-apc): respect mamba_ssm_cache_dtype + float32 intermediate states
- gated_delta_net_state_dtype now accepts mamba_ssm_cache_dtype and
  delegates to _mamba_state_dtype (matching Mamba1/Mamba2)
- Both call sites in qwen3_next.py pass mamba_ssm_cache_dtype through
- FLA h tensor allocated as float32 (matching final_state) to avoid
  precision loss in intermediate states used for APC block caching
- Output kernel casts b_h to b_q.dtype before tl.dot

+++ vllm/model_executor/layers/fla/ops/chunk_delta_h.py
@@ -312,7 +312,7 @@ def chunk_gated_delta_rule_fwd_h(
         )
     assert K <= 256, "current kernel does not support head dimension larger than 256."
 
-    h = k.new_empty(B, NT, H, V, K)
+    h = k.new_empty(B, NT, H, V, K, dtype=torch.float32)
     final_state = (
         k.new_empty(N, H, V, K, dtype=torch.float32) if output_final_state else None
     )
